phantom-guard-pip/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phantom-guard-pip"
version = "0.3.0"
description = "pip install wrapper with slopsquatting protection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Matteo", email = "noreply@example.com" }
]
keywords = [
    "security",
    "pip",
    "slopsquatting",
    "supply-chain",
    "phantom-guard"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "phantom-guard>=0.2.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]

[project.scripts]
phantom-pip = "phantom_pip.cli:app"

[project.urls]
Homepage = "https://github.com/matte1782/phantom_guard"
Repository = "https://github.com/matte1782/phantom_guard"
Documentation = "https://github.com/matte1782/phantom_guard#readme"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/phantom_pip"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/phantom_pip"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 90

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "SIM", "ARG"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
